+++ web/components/control/index.js
@@ -1,0 +1,6 @@
+import { FeatureSelector } from "./FeatureSelector";
+import { PlaylistSelector } from "./PlaylistSelector";
+export {
+    FeatureSelector,
+    PlaylistSelector
+}
